+++ crates/bevy_transform/src/systems.rs
@@ -1,5 +1,7 @@
 #[cfg(test)]
 mod test {
+    // TODO: move these tests to bevy_hierarchy::Propagate module
+
     use bevy_app::prelude::*;
     use bevy_ecs::prelude::*;
     use bevy_ecs::system::CommandQueue;
@@ -7,7 +9,7 @@ mod test {
 
     use crate::components::{GlobalTransform, Transform};
     use crate::TransformBundle;
-    use bevy_hierarchy::{BuildChildren, BuildWorldChildren, Children, Parent};
+    use bevy_hierarchy::{BuildChildren, BuildWorldChildren, Children};
 
     #[test]
     fn did_propagate() {
@@ -95,98 +97,6 @@ mod test {
     }
 
     #[test]
-    fn correct_children() {
-        let mut world = World::default();
-
-        let mut update_stage = SystemStage::parallel();
-        update_stage.add_system(bevy_hierarchy::propagate_system::<Transform>);
-
-        let mut schedule = Schedule::default();
-        schedule.add_stage("update", update_stage);
-
-        // Add parent entities
-        let mut children = Vec::new();
-        let parent = {
-            let mut command_queue = CommandQueue::default();
-            let mut commands = Commands::new(&mut command_queue, &world);
-            let parent = commands
-                .spawn()
-                .insert(Transform::from_xyz(1.0, 0.0, 0.0))
-                .id();
-            commands.entity(parent).with_children(|parent| {
-                children.push(
-                    parent
-                        .spawn()
-                        .insert(Transform::from_xyz(0.0, 2.0, 0.0))
-                        .id(),
-                );
-                children.push(
-                    parent
-                        .spawn()
-                        .insert(Transform::from_xyz(0.0, 3.0, 0.0))
-                        .id(),
-                );
-            });
-            command_queue.apply(&mut world);
-            schedule.run(&mut world);
-            parent
-        };
-
-        assert_eq!(
-            world
-                .get::<Children>(parent)
-                .unwrap()
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            children,
-        );
-
-        // Parent `e1` to `e2`.
-        {
-            let mut command_queue = CommandQueue::default();
-            let mut commands = Commands::new(&mut command_queue, &world);
-            commands.entity(children[1]).add_child(children[0]);
-            command_queue.apply(&mut world);
-            schedule.run(&mut world);
-        }
-
-        assert_eq!(
-            world
-                .get::<Children>(parent)
-                .unwrap()
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            vec![children[1]]
-        );
-
-        assert_eq!(
-            world
-                .get::<Children>(children[1])
-                .unwrap()
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            vec![children[0]]
-        );
-
-        assert!(world.despawn(children[0]));
-
-        schedule.run(&mut world);
-
-        assert_eq!(
-            world
-                .get::<Children>(parent)
-                .unwrap()
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            vec![children[1]]
-        );
-    }
-
-    #[test]
     fn correct_transforms_when_no_children() {
         let mut app = App::new();
 
@@ -226,48 +136,5 @@ mod test {
         for global in state.iter(&app.world) {
             assert_eq!(global, &GlobalTransform::from_translation(translation));
         }
-    }
-
-    #[test]
-    #[should_panic]
-    fn panic_when_hierarchy_cycle() {
-        // We cannot directly edit Parent and Children, so we use a temp world to break
-        // the hierarchy's invariants.
-        let mut temp = World::new();
-        let mut app = App::new();
-
-        app.add_system(bevy_hierarchy::propagate_system::<Transform>);
-
-        fn setup_world(world: &mut World) -> (Entity, Entity) {
-            let mut grandchild = Entity::from_raw(0);
-            let child = world
-                .spawn()
-                .insert_bundle((Transform::identity(), GlobalTransform::default()))
-                .with_children(|builder| {
-                    grandchild = builder
-                        .spawn()
-                        .insert_bundle((Transform::identity(), GlobalTransform::default()))
-                        .id();
-                })
-                .id();
-            (child, grandchild)
-        }
-
-        let (temp_child, temp_grandchild) = setup_world(&mut temp);
-        let (child, grandchild) = setup_world(&mut app.world);
-
-        assert_eq!(temp_child, child);
-        assert_eq!(temp_grandchild, grandchild);
-
-        app.world
-            .spawn()
-            .insert_bundle((Transform::default(), GlobalTransform::default()))
-            .push_children(&[child]);
-        std::mem::swap(
-            &mut *app.world.get_mut::<Parent>(child).unwrap(),
-            &mut *temp.get_mut::<Parent>(grandchild).unwrap(),
-        );
-
-        app.update();
     }
 }
